Fix Type Signatures

+++ c-allocators/src/float_array.h
@@ -12,31 +12,31 @@ struct FloatArray {
 // Clear all Elements from the Dynamic Float Array
 // Assigns the Value of all Elements to NULL
 // Returns the Array
-int clear_array(struct FloatArray initial_array);
+float clear_array(struct FloatArray initial_array);
 
 // Create Dynamic Array
 // Returns the Array
 // Returns NULL if Memory Allocation Failed
-int create_array(struct FloatArray initial_array);
+float create_array(struct FloatArray initial_array);
 
 // Insert an Element into the Dynamic Float Array
 // At an Index Position
 // Shifts all Elements after it, to the Right
 // Returns the Array
 // Returns NULL if Memory Allocation Failed
-int insert_element(struct FloatArray initial_array, int element, int index);
+float insert_element(struct FloatArray initial_array, int element, int index);
 
 // Append an Element to the End of the Dynamic Float Array
 // Returns the Array
 // Returns NULL if Memory Allocation Failed
-int push_element(struct FloatArray initial_array, int element);
+float push_element(struct FloatArray initial_array, int element);
 
 // Remove an Element from of the Dynamic Float Array
 // At an Index Position
 // Returns the Element
 // Shifts all Elements after it, to the Left
 // Returns NULL if Memory Allocation Failed
-int remove_element(struct FloatArray initial_array, int index);
+float  remove_element(struct FloatArray initial_array, int index);
 
 #endif
